Set all load balances to be represented as Longs instead of some being represented as Integers and some as Longs.
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@25159 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/org/starexec/data/database/Queues.java
@@ -659,7 +659,7 @@ public class Queues {
 	 * @return The integer sum of wallclock timeouts, or null on failure
 	 */
 
-	public static Integer getUserLoadOnQueue(int queueId, int userId) {
+	public static Long getUserLoadOnQueue(int queueId, int userId) {
 		Connection con = null;	
 		CallableStatement procedure = null;
 		ResultSet results = null;
@@ -671,7 +671,7 @@ public class Queues {
 			results = procedure.executeQuery();
 
 			while(results.next()){
-				return results.getInt("queue_load");	
+				return results.getLong("queue_load");
 			}							
 		} catch (Exception e){			
 			log.error(e.getMessage(), e);
